Get DW-nominate Scores for members of the U.S. Senate

Returns a dataset of nominate scores for Senate members of the the specified congress, available otherwise at VoteView.com.

get_senate_nominate(congress = "ALL")

Arguments

congress
Number for the congress you're interested in. Defaults to "ALL" to get the most recent scores for every congress.

Value

a data frame containing the relevant legislator names and DW-nominate scores, as well as otherinformation made available by VoteView

See also

Examples

 # get the scores for the 116th (2019-2021) congress
 senate <- get_senate_nominate(congress=116)

 # returned as a data frame
 senate

#>     congress chamber icpsr state_icpsr district_code state_abbrev party_code
#> 1        116  Senate 41705          41             0           AL        100
#> 2        116  Senate 94659          41             0           AL        200
#> 3        116  Senate 40300          81             0           AK        200
#> 4        116  Senate 41500          81             0           AK        200
#> 5        116  Senate 21300          61             0           AZ        100
#> 6        116  Senate 21501          61             0           AZ        200
#> 7        116  Senate 20101          42             0           AR        200
#> 8        116  Senate 21301          42             0           AR        200
#> 9        116  Senate 41701          71             0           CA        100
#> 10       116  Senate 49300          71             0           CA        100
#> 11       116  Senate 21112          62             0           CO        200
#> 12       116  Senate 40910          62             0           CO        100
#> 13       116  Senate 20707           1             0           CT        100
#> 14       116  Senate 41101           1             0           CT        100
#> 15       116  Senate 15015          11             0           DE        100
#> 16       116  Senate 40916          11             0           DE        100
#> 17       116  Senate 41102          43             0           FL        200
#> 18       116  Senate 41903          43             0           FL        200
#> 19       116  Senate 29909          44             0           GA        200
#> 20       116  Senate 41501          44             0           GA        200
#> 21       116  Senate 20713          82             0           HI        100
#> 22       116  Senate 41112          82             0           HI        100
#> 23       116  Senate 29345          63             0           ID        200
#> 24       116  Senate 40902          63             0           ID        200
#> 25       116  Senate 15021          21             0           IL        100
#> 26       116  Senate 21325          21             0           IL        100
#> 27       116  Senate 21133          22             0           IN        200
#> 28       116  Senate 41900          22             0           IN        200
#> 29       116  Senate 14226          31             0           IA        200
#> 30       116  Senate 41502          31             0           IA        200
#> 31       116  Senate 14852          32             0           KS        200
#> 32       116  Senate 29722          32             0           KS        200
#> 33       116  Senate 14921          51             0           KY        200
#> 34       116  Senate 41104          51             0           KY        200
#> 35       116  Senate 20919          45             0           LA        200
#> 36       116  Senate 41703          45             0           LA        200
#> 37       116  Senate 41300           2             0           ME        328
#> 38       116  Senate 49703           2             0           ME        200
#> 39       116  Senate 15408          52             0           MD        100
#> 40       116  Senate 20330          52             0           MD        100
#> 41       116  Senate 14435           3             0           MA        100
#> 42       116  Senate 41301           3             0           MA        100
#> 43       116  Senate 20923          23             0           MI        100
#> 44       116  Senate 29732          23             0           MI        100
#> 45       116  Senate 40700          33             0           MN        100
#> 46       116  Senate 41706          33             0           MN        100
#> 47       116  Senate 29534          46             0           MS        200
#> 48       116  Senate 41707          46             0           MS        200
#> 49       116  Senate 29735          34             0           MO        200
#> 50       116  Senate 41901          34             0           MO        200
#> 51       116  Senate 21338          64             0           MT        200
#> 52       116  Senate 40702          64             0           MT        100
#> 53       116  Senate 41302          35             0           NE        200
#> 54       116  Senate 41503          35             0           NE        200
#> 55       116  Senate 21743          65             0           NV        100
#> 56       116  Senate 41700          65             0           NV        100
#> 57       116  Senate 40906           4             0           NH        100
#> 58       116  Senate 41702           4             0           NH        100
#> 59       116  Senate 29373          12             0           NJ        100
#> 60       116  Senate 41308          12             0           NJ        100
#> 61       116  Senate 20930          66             0           NM        100
#> 62       116  Senate 29924          66             0           NM        100
#> 63       116  Senate 14858          13             0           NY        100
#> 64       116  Senate 20735          13             0           NY        100
#> 65       116  Senate 29548          47             0           NC        200
#> 66       116  Senate 41504          47             0           NC        200
#> 67       116  Senate 21350          36             0           ND        200
#> 68       116  Senate 41107          36             0           ND        200
#> 69       116  Senate 29386          24             0           OH        200
#> 70       116  Senate 29389          24             0           OH        100
#> 71       116  Senate 15424          53             0           OK        200
#> 72       116  Senate 21166          53             0           OK        200
#> 73       116  Senate 14871          72             0           OR        100
#> 74       116  Senate 40908          72             0           OR        100
#> 75       116  Senate 29935          14             0           PA        200
#> 76       116  Senate 40703          14             0           PA        100
#> 77       116  Senate 29142           5             0           RI        100
#> 78       116  Senate 40704           5             0           RI        100
#> 79       116  Senate 21173          48             0           SC        200
#> 80       116  Senate 29566          48             0           SC        200
#> 81       116  Senate 29754          37             0           SD        200
#> 82       116  Senate 41505          37             0           SD        200
#> 83       116  Senate 20351          54             0           TN        200
#> 84       116  Senate 40304          54             0           TN        200
#> 85       116  Senate 40305          49             0           TX        200
#> 86       116  Senate 41304          49             0           TX        200
#> 87       116  Senate 41110          67             0           UT        200
#> 88       116  Senate 41902          67             0           UT        200
#> 89       116  Senate 14307           6             0           VT        100
#> 90       116  Senate 29147           6             0           VT        328
#> 91       116  Senate 40909          40             0           VA        100
#> 92       116  Senate 41305          40             0           VA        100
#> 93       116  Senate 39310          73             0           WA        100
#> 94       116  Senate 49308          73             0           WA        100
#> 95       116  Senate 20146          56             0           WV        200
#> 96       116  Senate 40915          56             0           WV        100
#> 97       116  Senate 29940          25             0           WI        100
#> 98       116  Senate 41111          25             0           WI        200
#> 99       116  Senate 40707          68             0           WY        200
#> 100      116  Senate 49706          68             0           WY        200
#>     occupancy last_means                             bioname bioguide_id born
#> 1          NA         NA        JONES, Gordon Douglas (Doug)     J000300 1954
#> 2          NA         NA                  SHELBY, Richard C.     S000320 1934
#> 3          NA         NA                     MURKOWSKI, Lisa     M001153 1957
#> 4          NA         NA              SULLIVAN, Daniel Scott     S001198 1964
#> 5          NA         NA                     SINEMA, Kyrsten     S001191 1976
#> 6          NA         NA                     McSALLY, Martha     M001197 1966
#> 7          NA         NA                       BOOZMAN, John     B001236 1950
#> 8          NA         NA                         COTTON, Tom     C001095 1977
#> 9          NA         NA                 HARRIS, Kamala Devi     H001075 1964
#> 10         NA         NA                   FEINSTEIN, Dianne     F000062 1933
#> 11         NA         NA                       GARDNER, Cory     G000562 1974
#> 12         NA         NA                  BENNET, Michael F.     B001267 1964
#> 13         NA         NA                 MURPHY, Christopher     M001169 1973
#> 14         NA         NA                 BLUMENTHAL, Richard     B001277 1946
#> 15         NA         NA              CARPER, Thomas Richard     C000174 1947
#> 16         NA         NA               COONS, Christopher A.     C001088 1963
#> 17         NA         NA                        RUBIO, Marco     R000595 1971
#> 18         NA         NA          SCOTT, Richard Lynn (Rick)     S001217 1952
#> 19         NA         NA                     ISAKSON, Johnny     I000055 1944
#> 20         NA         NA           PERDUE, David Alfred, Jr.     P000612 1949
#> 21         NA         NA                       HIRONO, Mazie     H001042 1947
#> 22         NA         NA               SCHATZ, Brian Emanuel     S001194 1972
#> 23         NA         NA                 CRAPO, Michael Dean     C000880 1951
#> 24         NA         NA                        RISCH, James     R000584 1943
#> 25         NA         NA              DURBIN, Richard Joseph     D000563 1944
#> 26         NA         NA                    DUCKWORTH, Tammy     D000622 1968
#> 27         NA         NA                         YOUNG, Todd     Y000064 1972
#> 28         NA         NA                      BRAUN, Michael     B001310 1954
#> 29         NA         NA            GRASSLEY, Charles Ernest     G000386 1933
#> 30         NA         NA                         ERNST, Joni     E000295 1970
#> 31         NA         NA      ROBERTS, Charles Patrick (Pat)     R000307 1936
#> 32         NA         NA                        MORAN, Jerry     M000934 1954
#> 33         NA         NA McCONNELL, Addison Mitchell (Mitch)     M000355 1942
#> 34         NA         NA                          PAUL, Rand     P000603 1963
#> 35         NA         NA                       CASSIDY, Bill     C001075 1957
#> 36         NA         NA                 KENNEDY, John Neely     K000393 1951
#> 37         NA         NA            KING, Angus Stanley, Jr.     K000383 1944
#> 38         NA         NA             COLLINS, Susan Margaret     C001035 1952
#> 39         NA         NA              CARDIN, Benjamin Louis     C000141 1943
#> 40         NA         NA             VAN HOLLEN, Christopher     V000128 1959
#> 41         NA         NA                 MARKEY, Edward John     M000133 1946
#> 42         NA         NA                   WARREN, Elizabeth     W000817 1949
#> 43         NA         NA                     PETERS, Gary C.     P000595 1958
#> 44         NA         NA               STABENOW, Deborah Ann     S000770 1950
#> 45         NA         NA                      KLOBUCHAR, Amy     K000367 1960
#> 46         NA         NA                         SMITH, Tina     S001203 1958
#> 47         NA         NA                    WICKER, Roger F.     W000437 1951
#> 48         NA         NA                   HYDE-SMITH, Cindy     H001079 1959
#> 49         NA         NA                          BLUNT, Roy     B000575 1950
#> 50         NA         NA                HAWLEY, Joshua David     H001089 1979
#> 51         NA         NA                       DAINES, Steve     D000618 1962
#> 52         NA         NA                         TESTER, Jon     T000464 1956
#> 53         NA         NA                FISCHER, Debra (Deb)     F000463 1951
#> 54         NA         NA                SASSE, Benjamin Eric     S001197 1972
#> 55         NA         NA               ROSEN, Jacklyn Sheryl     R000608 1957
#> 56         NA         NA       CORTEZ MASTO, Catherine Marie     C001113 1964
#> 57         NA         NA                     SHAHEEN, Jeanne     S001181 1947
#> 58         NA         NA           HASSAN, Margaret (Maggie)     H001076 1958
#> 59         NA         NA                    MENENDEZ, Robert     M000639 1954
#> 60         NA         NA                BOOKER, Cory Anthony     B001288 1969
#> 61         NA         NA                    HEINRICH, Martin     H001046 1971
#> 62         NA         NA                 UDALL, Thomas (Tom)     U000039 1948
#> 63         NA         NA      SCHUMER, Charles Ellis (Chuck)     S000148 1950
#> 64         NA         NA                 GILLIBRAND, Kirsten     G000555 1966
#> 65         NA         NA                    BURR, Richard M.     B001135 1955
#> 66         NA         NA        TILLIS, Thomas Roland (Thom)     T000476 1960
#> 67         NA         NA                       CRAMER, Kevin     C001096 1961
#> 68         NA         NA                        HOEVEN, John     H001061 1957
#> 69         NA         NA         PORTMAN, Robert Jones (Rob)     P000449 1955
#> 70         NA         NA                      BROWN, Sherrod     B000944 1952
#> 71         NA         NA              INHOFE, James Mountain     I000024 1934
#> 72         NA         NA                     LANKFORD, James     L000575 1968
#> 73         NA         NA                   WYDEN, Ronald Lee     W000779 1949
#> 74         NA         NA                       MERKLEY, Jeff     M001176 1956
#> 75         NA         NA              TOOMEY, Patrick Joseph     T000461 1961
#> 76         NA         NA            CASEY, Robert (Bob), Jr.     C001070 1960
#> 77         NA         NA                REED, John F. (Jack)     R000122 1949
#> 78         NA         NA                 WHITEHOUSE, Sheldon     W000802 1955
#> 79         NA         NA                          SCOTT, Tim     S001184 1965
#> 80         NA         NA                  GRAHAM, Lindsey O.     G000359 1955
#> 81         NA         NA                         THUNE, John     T000250 1961
#> 82         NA         NA       ROUNDS, Marion Michael (Mike)     R000605 1954
#> 83         NA         NA                   BLACKBURN, Marsha     B001243 1952
#> 84         NA         NA                    ALEXANDER, Lamar     A000360 1940
#> 85         NA         NA                        CORNYN, John     C001056 1952
#> 86         NA         NA           CRUZ, Rafael Edward (Ted)     C001098 1970
#> 87         NA         NA                           LEE, Mike     L000577 1971
#> 88         NA         NA         ROMNEY, Willard Mitt (Mitt)     R000615 1947
#> 89         NA         NA               LEAHY, Patrick Joseph     L000174 1940
#> 90         NA         NA                    SANDERS, Bernard     S000033 1941
#> 91         NA         NA                        WARNER, Mark     W000805 1954
#> 92         NA         NA        KAINE, Timothy Michael (Tim)     K000384 1958
#> 93         NA         NA                  CANTWELL, Maria E.     C000127 1958
#> 94         NA         NA                       MURRAY, Patty     M001111 1950
#> 95         NA         NA               CAPITO, Shelley Moore     C001047 1953
#> 96         NA         NA                   MANCHIN, Joe, III     M001183 1947
#> 97         NA         NA                      BALDWIN, Tammy     B001230 1962
#> 98         NA         NA                        JOHNSON, Ron     J000293 1955
#> 99         NA         NA                   BARRASSO, John A.     B001261 1952
#> 100        NA         NA                    ENZI, Michael B.     E000285 1944
#>     died nominate_dim1 nominate_dim2 nominate_log_likelihood
#> 1     NA        -0.091         0.170               -31.58891
#> 2     NA         0.431         0.526               -22.63797
#> 3     NA         0.209        -0.287               -15.00029
#> 4     NA         0.469         0.084               -18.05335
#> 5     NA        -0.104         0.059               -35.88232
#> 6     NA         0.346         0.016               -19.73622
#> 7     NA         0.400         0.245               -12.10188
#> 8     NA         0.594         0.094               -33.08733
#> 9     NA        -0.713        -0.078               -10.80739
#> 10    NA        -0.269        -0.207               -31.80255
#> 11    NA         0.448        -0.018               -16.46835
#> 12    NA        -0.212        -0.115               -34.17949
#> 13    NA        -0.294        -0.181               -47.17134
#> 14    NA        -0.429        -0.100               -38.90498
#> 15    NA        -0.175        -0.225               -40.81844
#> 16    NA        -0.232        -0.211               -25.30205
#> 17    NA         0.582        -0.294               -21.76534
#> 18    NA         0.407         0.352                -8.08169
#> 19    NA         0.402        -0.011                -9.50031
#> 20    NA         0.579        -0.107               -12.80389
#> 21    NA        -0.501        -0.088               -31.64457
#> 22    NA        -0.435        -0.065               -34.96180
#> 23    NA         0.510         0.290               -10.94464
#> 24    NA         0.630         0.572                -8.57312
#> 25    NA        -0.356        -0.351               -51.23738
#> 26    NA        -0.330         0.108               -33.91876
#> 27    NA         0.481        -0.004               -38.18633
#> 28    NA         0.784         0.620               -11.58184
#> 29    NA         0.345        -0.068               -35.19229
#> 30    NA         0.503         0.045               -12.12663
#> 31    NA         0.414        -0.092                -8.33982
#> 32    NA         0.414         0.213               -50.29245
#> 33    NA         0.403         0.002               -23.33385
#> 34    NA         0.876        -0.482               -57.96170
#> 35    NA         0.452        -0.092               -11.06320
#> 36    NA         0.558         0.283               -29.96594
#> 37    NA        -0.154        -0.238               -25.91592
#> 38    NA         0.112        -0.549               -21.17594
#> 39    NA        -0.324        -0.239               -39.58995
#> 40    NA        -0.390        -0.185               -32.91680
#> 41    NA        -0.506        -0.440               -33.64069
#> 42    NA        -0.769        -0.277                -7.79274
#> 43    NA        -0.244        -0.221               -48.06622
#> 44    NA        -0.336         0.017               -46.23197
#> 45    NA        -0.269        -0.288               -98.61275
#> 46    NA        -0.389        -0.090               -29.40422
#> 47    NA         0.377         0.357               -10.16710
#> 48    NA         0.383         0.277                -7.63224
#> 49    NA         0.429         0.286               -10.04207
#> 50    NA         0.572        -0.310               -21.91135
#> 51    NA         0.546        -0.114               -28.75963
#> 52    NA        -0.213         0.126               -32.47864
#> 53    NA         0.467         0.320                -8.33917
#> 54    NA         0.802        -0.265               -17.42175
#> 55    NA        -0.258         0.242               -20.41785
#> 56    NA        -0.368         0.325               -26.06746
#> 57    NA        -0.246        -0.179               -34.52402
#> 58    NA        -0.253        -0.055               -26.94159
#> 59    NA        -0.365        -0.120               -48.81647
#> 60    NA        -0.607        -0.202               -18.68618
#> 61    NA        -0.307        -0.029               -37.74538
#> 62    NA        -0.453         0.171               -35.50266
#> 63    NA        -0.353        -0.390               -36.69194
#> 64    NA        -0.439        -0.303               -38.75699
#> 65    NA         0.452        -0.055                -8.43347
#> 66    NA         0.421         0.047               -12.83357
#> 67    NA         0.390         0.297                -7.61101
#> 68    NA         0.341         0.331                -8.77748
#> 69    NA         0.374        -0.243               -11.15851
#> 70    NA        -0.433        -0.126               -48.31060
#> 71    NA         0.554         0.041               -22.65606
#> 72    NA         0.586         0.155               -14.25792
#> 73    NA        -0.323        -0.442               -39.11528
#> 74    NA        -0.462        -0.774               -29.12795
#> 75    NA         0.645        -0.293               -18.57566
#> 76    NA        -0.307         0.192               -39.21195
#> 77    NA        -0.377        -0.212               -36.13636
#> 78    NA        -0.378        -0.145               -45.59589
#> 79    NA         0.644         0.007               -17.84925
#> 80    NA         0.406        -0.173               -13.46314
#> 81    NA         0.410         0.134                -8.65937
#> 82    NA         0.384         0.067                -7.00070
#> 83    NA         0.616         0.132               -14.63792
#> 84    NA         0.324        -0.177               -12.82492
#> 85    NA         0.494        -0.004               -10.57963
#> 86    NA         0.823        -0.312               -18.13274
#> 87    NA         0.916        -0.401               -55.45187
#> 88    NA         0.368         0.576               -25.04355
#> 89    NA        -0.366        -0.132               -44.56409
#> 90    NA        -0.526        -0.371               -29.70131
#> 91    NA        -0.199        -0.036               -23.96768
#> 92    NA        -0.241        -0.078               -38.19748
#> 93    NA        -0.292        -0.395               -35.25270
#> 94    NA        -0.345        -0.277               -31.96662
#> 95    NA         0.261         0.056                -8.69824
#> 96    NA        -0.055         0.446               -26.86967
#> 97    NA        -0.510        -0.215               -45.12841
#> 98    NA         0.602        -0.295               -17.47652
#> 99    NA         0.538         0.236               -13.92419
#> 100   NA         0.542         0.193               -14.42722
#>     nominate_geo_mean_probability nominate_number_of_votes
#> 1                         0.87903                      245
#> 2                         0.91242                      247
#> 3                         0.93686                      230
#> 4                         0.92725                      239
#> 5                         0.85613                      231
#> 6                         0.92260                      245
#> 7                         0.95219                      247
#> 8                         0.87463                      247
#> 9                         0.92622                      141
#> 10                        0.87919                      247
#> 11                        0.93447                      243
#> 12                        0.83047                      184
#> 13                        0.82615                      247
#> 14                        0.85262                      244
#> 15                        0.84653                      245
#> 16                        0.89954                      239
#> 17                        0.91330                      240
#> 18                        0.96702                      241
#> 19                        0.95338                      199
#> 20                        0.94539                      228
#> 21                        0.87351                      234
#> 22                        0.86752                      246
#> 23                        0.95596                      243
#> 24                        0.96547                      244
#> 25                        0.80848                      241
#> 26                        0.86398                      232
#> 27                        0.85234                      239
#> 28                        0.95401                      246
#> 29                        0.86721                      247
#> 30                        0.95093                      241
#> 31                        0.96667                      246
#> 32                        0.79481                      219
#> 33                        0.90986                      247
#> 34                        0.77724                      230
#> 35                        0.95181                      224
#> 36                        0.88353                      242
#> 37                        0.90039                      247
#> 38                        0.91784                      247
#> 39                        0.85135                      246
#> 40                        0.87476                      246
#> 41                        0.86556                      233
#> 42                        0.95740                      179
#> 43                        0.82316                      247
#> 44                        0.82610                      242
#> 45                        0.59833                      192
#> 46                        0.88777                      247
#> 47                        0.95967                      247
#> 48                        0.96920                      244
#> 49                        0.95984                      245
#> 50                        0.91511                      247
#> 51                        0.88751                      241
#> 52                        0.87679                      247
#> 53                        0.96599                      241
#> 54                        0.93026                      241
#> 55                        0.91973                      244
#> 56                        0.89867                      244
#> 57                        0.86956                      247
#> 58                        0.89666                      247
#> 59                        0.82067                      247
#> 60                        0.87162                      136
#> 61                        0.85220                      236
#> 62                        0.86407                      243
#> 63                        0.86196                      247
#> 64                        0.77230                      150
#> 65                        0.96337                      226
#> 66                        0.94793                      240
#> 67                        0.96916                      243
#> 68                        0.96467                      244
#> 69                        0.95583                      247
#> 70                        0.82037                      244
#> 71                        0.91027                      241
#> 72                        0.94391                      247
#> 73                        0.85075                      242
#> 74                        0.88876                      247
#> 75                        0.92461                      237
#> 76                        0.85321                      247
#> 77                        0.86390                      247
#> 78                        0.82892                      243
#> 79                        0.93001                      246
#> 80                        0.94610                      243
#> 81                        0.96555                      247
#> 82                        0.97002                      230
#> 83                        0.94200                      245
#> 84                        0.93992                      207
#> 85                        0.95757                      244
#> 86                        0.92604                      236
#> 87                        0.79522                      242
#> 88                        0.90358                      247
#> 89                        0.83369                      245
#> 90                        0.82036                      150
#> 91                        0.90420                      238
#> 92                        0.85230                      239
#> 93                        0.86598                      245
#> 94                        0.87578                      241
#> 95                        0.96381                      236
#> 96                        0.89532                      243
#> 97                        0.83239                      246
#> 98                        0.93115                      245
#> 99                        0.94519                      247
#> 100                       0.94281                      245
#>     nominate_number_of_errors conditional nokken_poole_dim1 nokken_poole_dim2
#> 1                           9          NA            -0.077             0.244
#> 2                          11          NA             0.530            -0.032
#> 3                           5          NA             0.311            -0.412
#> 4                           6          NA             0.503             0.016
#> 5                          10          NA            -0.059             0.203
#> 6                           8          NA             0.363             0.380
#> 7                           4          NA             0.322             0.181
#> 8                          13          NA             0.450            -0.110
#> 9                           5          NA            -0.773            -0.129
#> 10                         11          NA            -0.276            -0.077
#> 11                          8          NA             0.292             0.021
#> 12                         15          NA            -0.271             0.022
#> 13                         23          NA            -0.192            -0.319
#> 14                         21          NA            -0.458            -0.039
#> 15                         14          NA            -0.196            -0.327
#> 16                          9          NA            -0.194            -0.240
#> 17                          9          NA             0.591            -0.306
#> 18                          3          NA             0.405             0.348
#> 19                          2          NA             0.334             0.146
#> 20                          8          NA             0.458             0.312
#> 21                         14          NA            -0.574            -0.052
#> 22                         20          NA            -0.474            -0.039
#> 23                          5          NA             0.501             0.826
#> 24                          5          NA             0.502             0.865
#> 25                         27          NA            -0.269            -0.234
#> 26                         16          NA            -0.388             0.338
#> 27                         18          NA             0.520            -0.546
#> 28                          5          NA             0.779             0.627
#> 29                         13          NA             0.652             0.125
#> 30                          5          NA             0.420             0.285
#> 31                          0          NA             0.327             0.230
#> 32                         17          NA             0.301            -0.307
#> 33                          7          NA             0.368             0.056
#> 34                         31          NA             0.886            -0.463
#> 35                          3          NA             0.411             0.427
#> 36                         14          NA             0.706            -0.168
#> 37                          6          NA            -0.165            -0.264
#> 38                          4          NA             0.144            -0.457
#> 39                         17          NA            -0.284             0.017
#> 40                          9          NA            -0.362            -0.373
#> 41                         12          NA            -0.578            -0.701
#> 42                          3          NA            -0.808            -0.447
#> 43                         23          NA            -0.340            -0.237
#> 44                         21          NA            -0.406            -0.069
#> 45                         46          NA            -0.479            -0.549
#> 46                         15          NA            -0.422            -0.075
#> 47                          3          NA             0.410             0.291
#> 48                          2          NA             0.371             0.392
#> 49                          2          NA             0.368             0.236
#> 50                         10          NA             0.575            -0.305
#> 51                         18          NA             0.547            -0.269
#> 52                         12          NA            -0.228            -0.054
#> 53                          3          NA             0.428             0.395
#> 54                          7          NA             0.722            -0.289
#> 55                          2          NA            -0.288             0.293
#> 56                         15          NA            -0.319             0.277
#> 57                         17          NA            -0.232             0.006
#> 58                          9          NA            -0.244            -0.004
#> 59                         22          NA            -0.386            -0.124
#> 60                          9          NA            -0.588            -0.356
#> 61                         18          NA            -0.336            -0.116
#> 62                         15          NA            -0.431             0.014
#> 63                         18          NA            -0.398            -0.314
#> 64                         21          NA            -0.781            -0.625
#> 65                          2          NA             0.325             0.277
#> 66                          7          NA             0.392             0.346
#> 67                          3          NA             0.385             0.273
#> 68                          3          NA             0.370             0.326
#> 69                          2          NA             0.339             0.157
#> 70                         21          NA            -0.364            -0.068
#> 71                         15          NA             0.730            -0.170
#> 72                          8          NA             0.780             0.045
#> 73                         19          NA            -0.344            -0.603
#> 74                         13          NA            -0.418            -0.731
#> 75                          5          NA             0.876             0.141
#> 76                         20          NA            -0.325             0.234
#> 77                         22          NA            -0.345            -0.082
#> 78                         25          NA            -0.321             0.035
#> 79                          9          NA             0.510             0.319
#> 80                          4          NA             0.297            -0.035
#> 81                          3          NA             0.384             0.334
#> 82                          1          NA             0.344             0.320
#> 83                          6          NA             0.629             0.409
#> 84                          1          NA             0.314             0.091
#> 85                          3          NA             0.392             0.323
#> 86                          8          NA             0.849            -0.338
#> 87                         30          NA             0.870            -0.492
#> 88                         11          NA             0.341             0.534
#> 89                         23          NA            -0.284            -0.034
#> 90                         12          NA            -0.940             0.214
#> 91                          8          NA            -0.243            -0.025
#> 92                         15          NA            -0.268             0.033
#> 93                         19          NA            -0.293            -0.358
#> 94                         13          NA            -0.394            -0.027
#> 95                          0          NA             0.331             0.264
#> 96                          5          NA            -0.029             0.314
#> 97                         29          NA            -0.395            -0.183
#> 98                          6          NA             0.570             0.009
#> 99                         10          NA             0.671             0.263
#> 100                         5          NA             0.487             0.856