Get DW-nominate Scores for members of the U.S. Senate

Returns a dataset of nominate scores for Senate members of the the specified congress, available otherwise at VoteView.com.

get_senate_nominate(congress = "ALL")

Arguments

congress
Number for the congress you're interested in. Defaults to "ALL" to get the most recent scores for every congress.

Value

a data frame containing the relevant legislator names and DW-nominate scores, as well as otherinformation made available by VoteView

See also

Examples

 # get the scores for the 116th (2019-2021) congress
 senate <- get_senate_nominate(congress=116)

 # returned as a data frame
 senate

#>     congress chamber icpsr state_icpsr district_code state_abbrev party_code
#> 1        116  Senate 41705          41             0           AL        100
#> 2        116  Senate 94659          41             0           AL        200
#> 3        116  Senate 40300          81             0           AK        200
#> 4        116  Senate 41500          81             0           AK        200
#> 5        116  Senate 21300          61             0           AZ        100
#> 6        116  Senate 21501          61             0           AZ        200
#> 7        116  Senate 20101          42             0           AR        200
#> 8        116  Senate 21301          42             0           AR        200
#> 9        116  Senate 41701          71             0           CA        100
#> 10       116  Senate 49300          71             0           CA        100
#> 11       116  Senate 21112          62             0           CO        200
#> 12       116  Senate 40910          62             0           CO        100
#> 13       116  Senate 20707           1             0           CT        100
#> 14       116  Senate 41101           1             0           CT        100
#> 15       116  Senate 15015          11             0           DE        100
#> 16       116  Senate 40916          11             0           DE        100
#> 17       116  Senate 41102          43             0           FL        200
#> 18       116  Senate 41903          43             0           FL        200
#> 19       116  Senate 29909          44             0           GA        200
#> 20       116  Senate 41501          44             0           GA        200
#> 21       116  Senate 20713          82             0           HI        100
#> 22       116  Senate 41112          82             0           HI        100
#> 23       116  Senate 29345          63             0           ID        200
#> 24       116  Senate 40902          63             0           ID        200
#> 25       116  Senate 15021          21             0           IL        100
#> 26       116  Senate 21325          21             0           IL        100
#> 27       116  Senate 21133          22             0           IN        200
#> 28       116  Senate 41900          22             0           IN        200
#> 29       116  Senate 14226          31             0           IA        200
#> 30       116  Senate 41502          31             0           IA        200
#> 31       116  Senate 14852          32             0           KS        200
#> 32       116  Senate 29722          32             0           KS        200
#> 33       116  Senate 14921          51             0           KY        200
#> 34       116  Senate 41104          51             0           KY        200
#> 35       116  Senate 20919          45             0           LA        200
#> 36       116  Senate 41703          45             0           LA        200
#> 37       116  Senate 41300           2             0           ME        328
#> 38       116  Senate 49703           2             0           ME        200
#> 39       116  Senate 15408          52             0           MD        100
#> 40       116  Senate 20330          52             0           MD        100
#> 41       116  Senate 14435           3             0           MA        100
#> 42       116  Senate 41301           3             0           MA        100
#> 43       116  Senate 20923          23             0           MI        100
#> 44       116  Senate 29732          23             0           MI        100
#> 45       116  Senate 40700          33             0           MN        100
#> 46       116  Senate 41706          33             0           MN        100
#> 47       116  Senate 29534          46             0           MS        200
#> 48       116  Senate 41707          46             0           MS        200
#> 49       116  Senate 29735          34             0           MO        200
#> 50       116  Senate 41901          34             0           MO        200
#> 51       116  Senate 21338          64             0           MT        200
#> 52       116  Senate 40702          64             0           MT        100
#> 53       116  Senate 41302          35             0           NE        200
#> 54       116  Senate 41503          35             0           NE        200
#> 55       116  Senate 21743          65             0           NV        100
#> 56       116  Senate 41700          65             0           NV        100
#> 57       116  Senate 40906           4             0           NH        100
#> 58       116  Senate 41702           4             0           NH        100
#> 59       116  Senate 29373          12             0           NJ        100
#> 60       116  Senate 41308          12             0           NJ        100
#> 61       116  Senate 20930          66             0           NM        100
#> 62       116  Senate 29924          66             0           NM        100
#> 63       116  Senate 14858          13             0           NY        100
#> 64       116  Senate 20735          13             0           NY        100
#> 65       116  Senate 29548          47             0           NC        200
#> 66       116  Senate 41504          47             0           NC        200
#> 67       116  Senate 21350          36             0           ND        200
#> 68       116  Senate 41107          36             0           ND        200
#> 69       116  Senate 29386          24             0           OH        200
#> 70       116  Senate 29389          24             0           OH        100
#> 71       116  Senate 15424          53             0           OK        200
#> 72       116  Senate 21166          53             0           OK        200
#> 73       116  Senate 14871          72             0           OR        100
#> 74       116  Senate 40908          72             0           OR        100
#> 75       116  Senate 29935          14             0           PA        200
#> 76       116  Senate 40703          14             0           PA        100
#> 77       116  Senate 29142           5             0           RI        100
#> 78       116  Senate 40704           5             0           RI        100
#> 79       116  Senate 21173          48             0           SC        200
#> 80       116  Senate 29566          48             0           SC        200
#> 81       116  Senate 29754          37             0           SD        200
#> 82       116  Senate 41505          37             0           SD        200
#> 83       116  Senate 20351          54             0           TN        200
#> 84       116  Senate 40304          54             0           TN        200
#> 85       116  Senate 40305          49             0           TX        200
#> 86       116  Senate 41304          49             0           TX        200
#> 87       116  Senate 41110          67             0           UT        200
#> 88       116  Senate 41902          67             0           UT        200
#> 89       116  Senate 14307           6             0           VT        100
#> 90       116  Senate 29147           6             0           VT        328
#> 91       116  Senate 40909          40             0           VA        100
#> 92       116  Senate 41305          40             0           VA        100
#> 93       116  Senate 39310          73             0           WA        100
#> 94       116  Senate 49308          73             0           WA        100
#> 95       116  Senate 20146          56             0           WV        200
#> 96       116  Senate 40915          56             0           WV        100
#> 97       116  Senate 29940          25             0           WI        100
#> 98       116  Senate 41111          25             0           WI        200
#> 99       116  Senate 40707          68             0           WY        200
#> 100      116  Senate 49706          68             0           WY        200
#>     occupancy last_means                             bioname bioguide_id born
#> 1          NA         NA        JONES, Gordon Douglas (Doug)     J000300 1954
#> 2          NA         NA                  SHELBY, Richard C.     S000320 1934
#> 3          NA         NA                     MURKOWSKI, Lisa     M001153 1957
#> 4          NA         NA              SULLIVAN, Daniel Scott     S001198 1964
#> 5          NA         NA                     SINEMA, Kyrsten     S001191 1976
#> 6          NA         NA                     McSALLY, Martha     M001197 1966
#> 7          NA         NA                       BOOZMAN, John     B001236 1950
#> 8          NA         NA                         COTTON, Tom     C001095 1977
#> 9          NA         NA                 HARRIS, Kamala Devi     H001075 1964
#> 10         NA         NA                   FEINSTEIN, Dianne     F000062 1933
#> 11         NA         NA                       GARDNER, Cory     G000562 1974
#> 12         NA         NA                  BENNET, Michael F.     B001267 1964
#> 13         NA         NA                 MURPHY, Christopher     M001169 1973
#> 14         NA         NA                 BLUMENTHAL, Richard     B001277 1946
#> 15         NA         NA              CARPER, Thomas Richard     C000174 1947
#> 16         NA         NA               COONS, Christopher A.     C001088 1963
#> 17         NA         NA                        RUBIO, Marco     R000595 1971
#> 18         NA         NA          SCOTT, Richard Lynn (Rick)     S001217 1952
#> 19         NA         NA                     ISAKSON, Johnny     I000055 1944
#> 20         NA         NA           PERDUE, David Alfred, Jr.     P000612 1949
#> 21         NA         NA                       HIRONO, Mazie     H001042 1947
#> 22         NA         NA               SCHATZ, Brian Emanuel     S001194 1972
#> 23         NA         NA                 CRAPO, Michael Dean     C000880 1951
#> 24         NA         NA                        RISCH, James     R000584 1943
#> 25         NA         NA              DURBIN, Richard Joseph     D000563 1944
#> 26         NA         NA                    DUCKWORTH, Tammy     D000622 1968
#> 27         NA         NA                         YOUNG, Todd     Y000064 1972
#> 28         NA         NA                      BRAUN, Michael     B001310 1954
#> 29         NA         NA            GRASSLEY, Charles Ernest     G000386 1933
#> 30         NA         NA                         ERNST, Joni     E000295 1970
#> 31         NA         NA      ROBERTS, Charles Patrick (Pat)     R000307 1936
#> 32         NA         NA                        MORAN, Jerry     M000934 1954
#> 33         NA         NA McCONNELL, Addison Mitchell (Mitch)     M000355 1942
#> 34         NA         NA                          PAUL, Rand     P000603 1963
#> 35         NA         NA                       CASSIDY, Bill     C001075 1957
#> 36         NA         NA                 KENNEDY, John Neely     K000393 1951
#> 37         NA         NA            KING, Angus Stanley, Jr.     K000383 1944
#> 38         NA         NA             COLLINS, Susan Margaret     C001035 1952
#> 39         NA         NA              CARDIN, Benjamin Louis     C000141 1943
#> 40         NA         NA             VAN HOLLEN, Christopher     V000128 1959
#> 41         NA         NA                 MARKEY, Edward John     M000133 1946
#> 42         NA         NA                   WARREN, Elizabeth     W000817 1949
#> 43         NA         NA                     PETERS, Gary C.     P000595 1958
#> 44         NA         NA               STABENOW, Deborah Ann     S000770 1950
#> 45         NA         NA                      KLOBUCHAR, Amy     K000367 1960
#> 46         NA         NA                         SMITH, Tina     S001203 1958
#> 47         NA         NA                    WICKER, Roger F.     W000437 1951
#> 48         NA         NA                   HYDE-SMITH, Cindy     H001079 1959
#> 49         NA         NA                          BLUNT, Roy     B000575 1950
#> 50         NA         NA                HAWLEY, Joshua David     H001089 1979
#> 51         NA         NA                       DAINES, Steve     D000618 1962
#> 52         NA         NA                         TESTER, Jon     T000464 1956
#> 53         NA         NA                FISCHER, Debra (Deb)     F000463 1951
#> 54         NA         NA                SASSE, Benjamin Eric     S001197 1972
#> 55         NA         NA               ROSEN, Jacklyn Sheryl     R000608 1957
#> 56         NA         NA       CORTEZ MASTO, Catherine Marie     C001113 1964
#> 57         NA         NA                     SHAHEEN, Jeanne     S001181 1947
#> 58         NA         NA           HASSAN, Margaret (Maggie)     H001076 1958
#> 59         NA         NA                    MENENDEZ, Robert     M000639 1954
#> 60         NA         NA                BOOKER, Cory Anthony     B001288 1969
#> 61         NA         NA                    HEINRICH, Martin     H001046 1971
#> 62         NA         NA                 UDALL, Thomas (Tom)     U000039 1948
#> 63         NA         NA      SCHUMER, Charles Ellis (Chuck)     S000148 1950
#> 64         NA         NA                 GILLIBRAND, Kirsten     G000555 1966
#> 65         NA         NA                    BURR, Richard M.     B001135 1955
#> 66         NA         NA        TILLIS, Thomas Roland (Thom)     T000476 1960
#> 67         NA         NA                       CRAMER, Kevin     C001096 1961
#> 68         NA         NA                        HOEVEN, John     H001061 1957
#> 69         NA         NA         PORTMAN, Robert Jones (Rob)     P000449 1955
#> 70         NA         NA                      BROWN, Sherrod     B000944 1952
#> 71         NA         NA              INHOFE, James Mountain     I000024 1934
#> 72         NA         NA                     LANKFORD, James     L000575 1968
#> 73         NA         NA                   WYDEN, Ronald Lee     W000779 1949
#> 74         NA         NA                       MERKLEY, Jeff     M001176 1956
#> 75         NA         NA              TOOMEY, Patrick Joseph     T000461 1961
#> 76         NA         NA            CASEY, Robert (Bob), Jr.     C001070 1960
#> 77         NA         NA                REED, John F. (Jack)     R000122 1949
#> 78         NA         NA                 WHITEHOUSE, Sheldon     W000802 1955
#> 79         NA         NA                          SCOTT, Tim     S001184 1965
#> 80         NA         NA                  GRAHAM, Lindsey O.     G000359 1955
#> 81         NA         NA                         THUNE, John     T000250 1961
#> 82         NA         NA       ROUNDS, Marion Michael (Mike)     R000605 1954
#> 83         NA         NA                   BLACKBURN, Marsha     B001243 1952
#> 84         NA         NA                    ALEXANDER, Lamar     A000360 1940
#> 85         NA         NA                        CORNYN, John     C001056 1952
#> 86         NA         NA           CRUZ, Rafael Edward (Ted)     C001098 1970
#> 87         NA         NA                           LEE, Mike     L000577 1971
#> 88         NA         NA         ROMNEY, Willard Mitt (Mitt)     R000615 1947
#> 89         NA         NA               LEAHY, Patrick Joseph     L000174 1940
#> 90         NA         NA                    SANDERS, Bernard     S000033 1941
#> 91         NA         NA                        WARNER, Mark     W000805 1954
#> 92         NA         NA        KAINE, Timothy Michael (Tim)     K000384 1958
#> 93         NA         NA                  CANTWELL, Maria E.     C000127 1958
#> 94         NA         NA                       MURRAY, Patty     M001111 1950
#> 95         NA         NA               CAPITO, Shelley Moore     C001047 1953
#> 96         NA         NA                   MANCHIN, Joe, III     M001183 1947
#> 97         NA         NA                      BALDWIN, Tammy     B001230 1962
#> 98         NA         NA                        JOHNSON, Ron     J000293 1955
#> 99         NA         NA                   BARRASSO, John A.     B001261 1952
#> 100        NA         NA                    ENZI, Michael B.     E000285 1944
#>     died nominate_dim1 nominate_dim2 nominate_log_likelihood
#> 1     NA        -0.092         0.147               -26.32649
#> 2     NA         0.432         0.529               -20.60845
#> 3     NA         0.209        -0.285               -13.35411
#> 4     NA         0.473         0.086               -16.51655
#> 5     NA        -0.105         0.059               -29.78365
#> 6     NA         0.346         0.014               -18.60237
#> 7     NA         0.400         0.245               -10.87705
#> 8     NA         0.596         0.097               -30.27507
#> 9     NA        -0.713        -0.078               -10.80739
#> 10    NA        -0.270        -0.207               -29.97324
#> 11    NA         0.448        -0.019               -14.03547
#> 12    NA        -0.212        -0.119               -28.48341
#> 13    NA        -0.296        -0.178               -41.07575
#> 14    NA        -0.426        -0.099               -36.16375
#> 15    NA        -0.175        -0.225               -32.13927
#> 16    NA        -0.233        -0.208               -22.97330
#> 17    NA         0.579        -0.298               -19.33372
#> 18    NA         0.405         0.336                -5.84112
#> 19    NA         0.402        -0.011                -9.50031
#> 20    NA         0.585        -0.104               -10.20191
#> 21    NA        -0.499        -0.089               -29.83566
#> 22    NA        -0.435        -0.063               -33.25134
#> 23    NA         0.510         0.290                -9.06185
#> 24    NA         0.629         0.571                -7.93240
#> 25    NA        -0.358        -0.352               -46.24065
#> 26    NA        -0.329         0.108               -32.26915
#> 27    NA         0.482         0.007               -28.74642
#> 28    NA         0.779         0.627               -11.29983
#> 29    NA         0.345        -0.068               -33.89684
#> 30    NA         0.502         0.041                -9.94967
#> 31    NA         0.414        -0.092                -7.03125
#> 32    NA         0.414         0.219               -38.86829
#> 33    NA         0.403         0.002               -22.11848
#> 34    NA         0.876        -0.482               -57.63780
#> 35    NA         0.452        -0.092                -9.20979
#> 36    NA         0.552         0.275               -29.00564
#> 37    NA        -0.153        -0.234               -22.46634
#> 38    NA         0.112        -0.549               -19.66127
#> 39    NA        -0.324        -0.239               -37.38866
#> 40    NA        -0.390        -0.185               -31.38123
#> 41    NA        -0.506        -0.440               -32.32622
#> 42    NA        -0.769        -0.277                -7.79274
#> 43    NA        -0.244        -0.221               -46.38995
#> 44    NA        -0.335         0.019               -42.01214
#> 45    NA        -0.269        -0.291               -92.50924
#> 46    NA        -0.388        -0.087               -27.87506
#> 47    NA         0.377         0.358                -8.98820
#> 48    NA         0.389         0.278                -6.49720
#> 49    NA         0.429         0.286                -8.52765
#> 50    NA         0.574        -0.333               -19.21810
#> 51    NA         0.544        -0.118               -26.34694
#> 52    NA        -0.212         0.127               -24.43121
#> 53    NA         0.464         0.316                -6.12520
#> 54    NA         0.801        -0.267               -16.91705
#> 55    NA        -0.256         0.245               -17.91497
#> 56    NA        -0.373         0.332               -23.19225
#> 57    NA        -0.248        -0.179               -28.37277
#> 58    NA        -0.259        -0.031               -21.68782
#> 59    NA        -0.364        -0.121               -43.93697
#> 60    NA        -0.607        -0.202               -18.68618
#> 61    NA        -0.305        -0.029               -34.18020
#> 62    NA        -0.453         0.172               -33.56877
#> 63    NA        -0.353        -0.389               -34.41925
#> 64    NA        -0.439        -0.303               -38.67575
#> 65    NA         0.452        -0.055                -7.00000
#> 66    NA         0.420         0.042               -10.26318
#> 67    NA         0.391         0.298                -6.40699
#> 68    NA         0.341         0.332                -7.78966
#> 69    NA         0.374        -0.244                -9.55666
#> 70    NA        -0.432        -0.125               -45.48712
#> 71    NA         0.555         0.043               -19.68190
#> 72    NA         0.585         0.153               -13.16677
#> 73    NA        -0.322        -0.440               -33.99442
#> 74    NA        -0.466        -0.776               -26.61039
#> 75    NA         0.645        -0.294               -17.07687
#> 76    NA        -0.306         0.186               -33.37197
#> 77    NA        -0.377        -0.212               -33.78418
#> 78    NA        -0.379        -0.153               -40.54647
#> 79    NA         0.643         0.006               -16.84681
#> 80    NA         0.407        -0.172               -11.42153
#> 81    NA         0.411         0.134                -7.42493
#> 82    NA         0.385         0.063                -5.89705
#> 83    NA         0.615         0.132               -13.66798
#> 84    NA         0.323        -0.178               -11.14127
#> 85    NA         0.494        -0.006                -8.25821
#> 86    NA         0.822        -0.314               -17.67338
#> 87    NA         0.917        -0.399               -48.99890
#> 88    NA         0.339         0.533               -21.56337
#> 89    NA        -0.367        -0.131               -40.06677
#> 90    NA        -0.526        -0.371               -29.70131
#> 91    NA        -0.199        -0.034               -22.18249
#> 92    NA        -0.241        -0.086               -35.03843
#> 93    NA        -0.291        -0.393               -30.89476
#> 94    NA        -0.344        -0.275               -29.12514
#> 95    NA         0.261         0.055                -7.26685
#> 96    NA        -0.055         0.446               -20.63154
#> 97    NA        -0.511        -0.215               -42.08192
#> 98    NA         0.600        -0.299               -15.38271
#> 99    NA         0.538         0.236               -11.95863
#> 100   NA         0.541         0.192               -12.93657
#>     nominate_geo_mean_probability nominate_number_of_votes
#> 1                         0.88575                      217
#> 2                         0.91019                      219
#> 3                         0.93603                      202
#> 4                         0.92638                      216
#> 5                         0.86354                      203
#> 6                         0.91785                      217
#> 7                         0.95155                      219
#> 8                         0.87089                      219
#> 9                         0.92622                      141
#> 10                        0.87209                      219
#> 11                        0.93737                      217
#> 12                        0.85513                      182
#> 13                        0.82898                      219
#> 14                        0.84584                      216
#> 15                        0.86234                      217
#> 16                        0.89998                      218
#> 17                        0.91284                      212
#> 18                        0.97295                      213
#> 19                        0.95338                      199
#> 20                        0.95145                      205
#> 21                        0.86517                      206
#> 22                        0.85853                      218
#> 23                        0.95910                      217
#> 24                        0.96394                      216
#> 25                        0.80486                      213
#> 26                        0.85693                      209
#> 27                        0.87263                      211
#> 28                        0.94949                      218
#> 29                        0.85660                      219
#> 30                        0.95436                      213
#> 31                        0.96826                      218
#> 32                        0.81587                      191
#> 33                        0.90394                      219
#> 34                        0.75998                      210
#> 35                        0.95749                      212
#> 36                        0.87324                      214
#> 37                        0.90250                      219
#> 38                        0.91413                      219
#> 39                        0.84239                      218
#> 40                        0.86593                      218
#> 41                        0.86040                      215
#> 42                        0.95740                      179
#> 43                        0.80911                      219
#> 44                        0.82175                      214
#> 45                        0.61295                      189
#> 46                        0.88048                      219
#> 47                        0.95979                      219
#> 48                        0.97037                      216
#> 49                        0.96146                      217
#> 50                        0.91599                      219
#> 51                        0.88365                      213
#> 52                        0.89444                      219
#> 53                        0.97165                      213
#> 54                        0.92433                      215
#> 55                        0.92041                      216
#> 56                        0.89819                      216
#> 57                        0.87849                      219
#> 58                        0.90571                      219
#> 59                        0.81822                      219
#> 60                        0.86800                      132
#> 61                        0.84846                      208
#> 62                        0.85544                      215
#> 63                        0.85456                      219
#> 64                        0.77138                      149
#> 65                        0.96643                      205
#> 66                        0.95274                      212
#> 67                        0.97064                      215
#> 68                        0.96458                      216
#> 69                        0.95730                      219
#> 70                        0.81011                      216
#> 71                        0.91174                      213
#> 72                        0.94165                      219
#> 73                        0.85312                      214
#> 74                        0.88558                      219
#> 75                        0.92154                      209
#> 76                        0.85866                      219
#> 77                        0.85704                      219
#> 78                        0.82813                      215
#> 79                        0.92563                      218
#> 80                        0.94873                      217
#> 81                        0.96666                      219
#> 82                        0.97123                      202
#> 83                        0.93896                      217
#> 84                        0.93966                      179
#> 85                        0.96249                      216
#> 86                        0.91929                      210
#> 87                        0.79535                      214
#> 88                        0.90623                      219
#> 89                        0.83140                      217
#> 90                        0.82036                      150
#> 91                        0.90325                      218
#> 92                        0.85215                      219
#> 93                        0.86730                      217
#> 94                        0.87220                      213
#> 95                        0.96567                      208
#> 96                        0.90850                      215
#> 97                        0.82445                      218
#> 98                        0.93157                      217
#> 99                        0.94686                      219
#> 100                       0.94213                      217
#>     nominate_number_of_errors conditional nokken_poole_dim1 nokken_poole_dim2
#> 1                           7          NA            -0.076             0.204
#> 2                          10          NA             0.531            -0.057
#> 3                           5          NA             0.322            -0.378
#> 4                           6          NA             0.503             0.016
#> 5                          10          NA            -0.066             0.188
#> 6                           8          NA             0.361             0.373
#> 7                           4          NA             0.318             0.158
#> 8                          13          NA             0.450            -0.142
#> 9                           5          NA            -0.773            -0.129
#> 10                         11          NA            -0.280            -0.072
#> 11                          6          NA             0.289             0.009
#> 12                         13          NA            -0.271             0.022
#> 13                         21          NA            -0.197            -0.292
#> 14                         19          NA            -0.450            -0.039
#> 15                         12          NA            -0.189            -0.331
#> 16                          8          NA            -0.196            -0.219
#> 17                          8          NA             0.590            -0.332
#> 18                          1          NA             0.405             0.337
#> 19                          2          NA             0.334             0.146
#> 20                          5          NA             0.458             0.312
#> 21                         13          NA            -0.567            -0.060
#> 22                         19          NA            -0.476            -0.031
#> 23                          3          NA             0.501             0.827
#> 24                          5          NA             0.501             0.865
#> 25                         24          NA            -0.275            -0.234
#> 26                         15          NA            -0.391             0.347
#> 27                         14          NA             0.530            -0.475
#> 28                          5          NA             0.779             0.627
#> 29                         13          NA             0.652             0.115
#> 30                          4          NA             0.420             0.271
#> 31                          0          NA             0.323             0.206
#> 32                         13          NA             0.318            -0.289
#> 33                          7          NA             0.364             0.035
#> 34                         31          NA             0.886            -0.463
#> 35                          2          NA             0.411             0.423
#> 36                         14          NA             0.706            -0.187
#> 37                          6          NA            -0.155            -0.254
#> 38                          4          NA             0.144            -0.451
#> 39                         17          NA            -0.267             0.007
#> 40                          9          NA            -0.358            -0.379
#> 41                         11          NA            -0.572            -0.703
#> 42                          3          NA            -0.808            -0.447
#> 43                         23          NA            -0.362            -0.215
#> 44                         19          NA            -0.400            -0.067
#> 45                         44          NA            -0.479            -0.549
#> 46                         14          NA            -0.425            -0.068
#> 47                          3          NA             0.410             0.280
#> 48                          2          NA             0.369             0.383
#> 49                          2          NA             0.367             0.218
#> 50                          8          NA             0.574            -0.333
#> 51                         17          NA             0.547            -0.293
#> 52                         10          NA            -0.223            -0.039
#> 53                          1          NA             0.428             0.387
#> 54                          7          NA             0.723            -0.301
#> 55                          2          NA            -0.287             0.320
#> 56                         13          NA            -0.325             0.300
#> 57                         15          NA            -0.248             0.061
#> 58                          8          NA            -0.260             0.077
#> 59                         21          NA            -0.367            -0.165
#> 60                          9          NA            -0.588            -0.356
#> 61                         15          NA            -0.323            -0.125
#> 62                         15          NA            -0.434             0.022
#> 63                         17          NA            -0.392            -0.319
#> 64                         21          NA            -0.781            -0.625
#> 65                          2          NA             0.321             0.250
#> 66                          5          NA             0.392             0.334
#> 67                          3          NA             0.384             0.256
#> 68                          3          NA             0.368             0.313
#> 69                          2          NA             0.335             0.131
#> 70                         19          NA            -0.341            -0.065
#> 71                         12          NA             0.731            -0.196
#> 72                          8          NA             0.781             0.030
#> 73                         17          NA            -0.326            -0.615
#> 74                         12          NA            -0.426            -0.731
#> 75                          5          NA             0.877             0.122
#> 76                         16          NA            -0.317             0.174
#> 77                         21          NA            -0.343            -0.089
#> 78                         24          NA            -0.314            -0.002
#> 79                          9          NA             0.510             0.311
#> 80                          3          NA             0.300             0.001
#> 81                          3          NA             0.383             0.323
#> 82                          1          NA             0.341             0.299
#> 83                          6          NA             0.629             0.405
#> 84                          1          NA             0.308             0.057
#> 85                          2          NA             0.391             0.310
#> 86                          8          NA             0.851            -0.353
#> 87                         27          NA             0.900            -0.437
#> 88                         10          NA             0.339             0.534
#> 89                         20          NA            -0.290            -0.027
#> 90                         12          NA            -0.940             0.214
#> 91                          8          NA            -0.240            -0.022
#> 92                         14          NA            -0.271            -0.003
#> 93                         17          NA            -0.273            -0.362
#> 94                         12          NA            -0.385            -0.024
#> 95                          0          NA             0.327             0.240
#> 96                          3          NA            -0.030             0.316
#> 97                         27          NA            -0.394            -0.181
#> 98                          6          NA             0.571            -0.013
#> 99                          8          NA             0.671             0.257
#> 100                         5          NA             0.486             0.859